create inference base class
